% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_function_names.R
\name{get_function_names}
\alias{get_function_names}
\title{Trouver les noms de fonctions}
\usage{
get_function_names(file)
}
\arguments{
\item{file}{character Path to an R file (Rmd, qmd, R)}
}
\value{
vector of strings containing user function names
}
\description{
Cette fonction recherche tous les noms de fonctions créées par l'utilisateur
}
\examples{
working_files <- get_working_files()
if (length(working_files) > 0) {
  get_function_names(get_working_files()[[1]])
}
}
