% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_working_files.R
\name{get_working_files}
\alias{get_working_files}
\title{Trouver les fichiers de travail R}
\usage{
get_working_files(path = ".", recursive = TRUE)
}
\arguments{
\item{path}{character giving the root searching path}

\item{recursive}{logical telling if the search shoud be recursive or not}
}
\value{
vector of strings consisting of relative paths to the working R files
}
\description{
Cette fonction recherche tous les fichiers R susceptibles de contenir des définitions de fonction
}
\examples{
get_working_files()
}
