% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_con.R
\name{my_con}
\alias{my_con}
\title{Create PostgreSQL database connection}
\usage{
my_con()
}
\value{
S4 object PqConnection
}
\description{
This function creates the connection to the PostgreSQL database
}
\examples{
con <- my_con()
}
