% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_function_names.R
\name{get_function_names}
\alias{get_function_names}
\title{Trouver les noms de fonctions}
\usage{
get_function_names(vector_of_files)
}
\arguments{
\item{vector_of_files}{vector of character Path(s) to R file(s) (Rmd, qmd, R)}
}
\value{
data.frame with each function name associated with filepath of file
in which the function  definitiof strings containing user function names
}
\description{
Cette fonction recherche tous les noms de fonctions créées par l'utilisateur
et renvoie un dataframe renseignant
\itemize{
\item le nom de chaque fonction (colonne fonction)
\item le nom de chaque chemin de fichier associé à la fonction (colonne fichier)
}
}
\examples{
get_function_names(get_working_files())
}
